Moved markdown component to .stories
